+++ plugins/systems/input/CMakeLists.txt
@@ -1,8 +1,8 @@
 # Add an executable for the server project
 add_library(Input SHARED
     Input.cpp
-    ../../../Actions.cpp
-    ../../../Keys.cpp
+    ./Actions.cpp
+    ./Keys.cpp
 )
 
 target_link_libraries(Input
