+++ plugins/systems/sfml/CMakeLists.txt
@@ -3,8 +3,8 @@ add_library(Sfml SHARED
     Sfml.cpp
     Display.cpp
     Input.cpp
-    ../../../Keys.cpp
-    ../../../Actions.cpp
+    ../input/Keys.cpp
+    ../input/Actions.cpp
 )
 
 CPMAddPackage(
